branch-status: use int32 for commits ahead/behind
uint64 was overkill for divergence counts. Switch BranchStatusEntry.CommitsAhead
and CommitsBehind to int32 and marshal them directly (BSON int32). Document the
command in docs/COMMANDS.md and reflect the int32 type in docs/verify/branch-status.md.

+++ internal/backends/dolt/branch_status_test.go
@@ -92,7 +92,7 @@ func branchStatusMap(t *testing.T, b *Backend, dbName, base string, targets ...s
 	return res, m
 }
 
-func assertAheadBehind(t *testing.T, m map[string]backends.BranchStatusEntry, target string, ahead, behind uint64) {
+func assertAheadBehind(t *testing.T, m map[string]backends.BranchStatusEntry, target string, ahead, behind int32) {
 	t.Helper()
 	e, ok := m[target]
 	if !ok {
